// src/app/model/empleados.ts
export class Empleados {
    empleadoId: number;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    apellidoCasada: string;
    correo: string;
    telefono: string;
    celular: string;
    fechaIngreso: string;
    activo: string;
    comentario: string;
}